+++ web/embed.go
@@ -1,0 +1,6 @@
+package webstatic
+
+import "embed"
+
+//go:embed index.html app.js style.css
+var FS embed.FS